Add roller piece differentiation shell
