+++ src/main/java/frc/robot/subsystems/Superstructure.java
@@ -1,0 +1,20 @@
+package frc.robot.subsystems;
+
+import edu.wpi.first.wpilibj2.command.SubsystemBase;
+
+import frc.robot.subsystems.RollerSubsystem.HeldPiece;
+
+public class Superstructure extends SubsystemBase {
+    private final RollerSubsystem rollerSubsystem;
+    private final TiltedElevatorSubsystem tiltedElevatorSubsystem;
+
+    public Superstructure(RollerSubsystem rollerSubsystem, TiltedElevatorSubsystem tiltedElevatorSubsystem) {
+        this.rollerSubsystem = rollerSubsystem;
+        this.tiltedElevatorSubsystem = tiltedElevatorSubsystem;
+    }
+
+    @Override
+    public void periodic() {
+        tiltedElevatorSubsystem.pieceGrabbed = rollerSubsystem.getPiece() != HeldPiece.EMPTY;
+    }
+}
